+++ prisma/migrations/20260627100733_part_c_byok_key/migration.sql
@@ -1,4 +1,0 @@
--- AlterTable
-ALTER TABLE "User" ADD COLUMN "anthropicKeyEnc" TEXT;
-ALTER TABLE "User" ADD COLUMN "anthropicKeyLast4" TEXT;
-ALTER TABLE "User" ADD COLUMN "anthropicKeyUpdatedAt" DATETIME;